requirements.txt
presidio-analyzer
presidio-anonymizer
presidio-image-redactor
pytesseract
pillow
opencv-python
PyMuPDF
easyocr
spacy
python-dotenv
